Add files via upload
Regression is an important machine learning model for these kinds of problems. Predicting sales
of a company needs time series data of that company and based on that data the model can
predict the future sales of that company or product. So, in this research project we will analyze
the time series sales data of a company and will predict the sales of the company for the coming
quarter and for a specific product.
For this kind of project of sales predict, we will apply the linear regression and logistic
regression and evaluate the result based on the training, testing and validation set of the data.